Crate arg_enum_proc_macro

arg_enum_proc_macro

This crate consists in a procedural macro derive that provides the same implementations that clap the clap::arg_enum macro provides: std::fmt::Display, std::str::FromStr and a variants() function.

By using a procedural macro it allows documenting the enum fields correctly and avoids the requirement of expanding the macro to use the structure with cbindgen.

Derive Macros

ArgEnum

Implement std::fmt::Display, std::str::FromStr and variants().